app/requirements.txt
pynacl>=1.5.0
requests>=2.32.3
pandas>=2.2.3